+++ neberus-core/src/main/java/net/oneandone/neberus/annotation/ApiAllowedValues.java
@@ -8,15 +8,29 @@ import java.lang.annotation.Target;
 /**
  * Define the allowed values for a parameter. This will be used to replace placeholders in the example curl.
  */
-@Target({ ElementType.PARAMETER })
+@Target({ ElementType.PARAMETER, ElementType.FIELD, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ApiAllowedValues {
 
     /**
-     * Name or describe the allowed values for this parameter.
+     * The exact allowed values for this parameter.
      *
      * @return the value
      */
-    String value() default "";
+    String[] value() default {};
+
+    /**
+     * Hint about the allowed values.
+     *
+     * @return the valueHint
+     */
+    String valueHint() default "";
+
+    /**
+     * Allow all values of the specified enum.
+     *
+     * @return the enumClass
+     */
+    Class<? extends Enum> enumValues() default Enum.class;
 
 }